feat: license builder
