+++ internal/licenses/scanner.go
@@ -3,43 +3,38 @@ package licenses
 import (
 	"context"
 	"fmt"
-	"github.com/anchore/syft/syft/file"
-	"github.com/anchore/syft/syft/pkg"
 	"io"
 
-	"github.com/google/licensecheck"
-
 	"github.com/anchore/syft/internal/log"
+	"github.com/google/licensecheck"
 )
 
 const (
-	DefaultCoverageThreshold     = 75 // determined by experimentation
-	DefaultIncludeLicenseContent = false
+	DefaultCoverageThreshold = 75 // determined by experimentation
 )
 
 type Evidence struct {
-	//...
+	ID    string            // License identifier. (See licenses/README.md.)
+	Type  licensecheck.Type // The type of the license: BSD, MIT, etc.
+	Start int               // Start offset of match in text; match is at text[Start:End].
+	End   int               // End offset of match in text.
+	IsURL bool              // Whether match is a URL.
 }
 
 type Scanner interface {
-	IdentifyLicenseIDs(context.Context, io.Reader) ([]string, []byte, error)
-	//FileSearch(context.Context, file.LocationReadCloser) ([]file.License, error)
-	//PkgSearch(context.Context, file.LocationReadCloser) ([]pkg.License, error)
-	Search(context.Context, file.LocationReadCloser) ([]Evidence, error)
+	FindEvidence(context.Context, io.Reader) ([]Evidence, []byte, error)
 }
 
 var _ Scanner = (*scanner)(nil)
 
 type scanner struct {
-	coverageThreshold     float64 // between 0 and 100
-	includeLicenseContent bool
-	scanner               func([]byte) licensecheck.Coverage
+	coverageThreshold float64 // between 0 and 100
+	scanner           func([]byte) licensecheck.Coverage
 }
 
 type ScannerConfig struct {
-	CoverageThreshold     float64
-	IncludeLicenseContent bool
-	Scanner               func([]byte) licensecheck.Coverage
+	CoverageThreshold float64
+	Scanner           func([]byte) licensecheck.Coverage
 }
 
 type Option func(*scanner)
@@ -50,12 +45,6 @@ func WithCoverage(coverage float64) Option {
 	}
 }
 
-func WithIncludeLicenseContent(includeLicenseContent bool) Option {
-	return func(s *scanner) {
-		s.includeLicenseContent = includeLicenseContent
-	}
-}
-
 // NewDefaultScanner returns a scanner that uses a new instance of the default licensecheck package scanner.
 func NewDefaultScanner(o ...Option) (Scanner, error) {
 	s, err := licensecheck.NewScanner(licensecheck.BuiltinLicenses())
@@ -63,10 +52,10 @@ func NewDefaultScanner(o ...Option) (Scanner, error) {
 		log.WithFields("error", err).Trace("unable to create default license scanner")
 		return nil, fmt.Errorf("unable to create default license scanner: %w", err)
 	}
+
 	newScanner := &scanner{
-		coverageThreshold:     DefaultCoverageThreshold,
-		includeLicenseContent: DefaultIncludeLicenseContent,
-		scanner:               s.Scan,
+		coverageThreshold: DefaultCoverageThreshold,
+		scanner:           s.Scan,
 	}
 
 	for _, opt := range o {
@@ -83,8 +72,7 @@ func NewScanner(c *ScannerConfig) (Scanner, error) {
 	}
 
 	return &scanner{
-		coverageThreshold:     c.CoverageThreshold,
-		includeLicenseContent: c.IncludeLicenseContent,
-		scanner:               c.Scanner,
+		coverageThreshold: c.CoverageThreshold,
+		scanner:           c.Scanner,
 	}, nil
 }
