--- internal/licenses/test-fixtures/Knuth-CTAN
@@ -1,5 +1,0 @@
-Knuth CTAN License
-Full name
-Knuth CTAN License
-
-This software is copyrighted. Unlimited copying and redistribution of this package and/or its individual files are permitted as long as there are no modifications. Modifications, and redistribution of modifications, are also permitted, but only if the resulting package and/or files are renamed.
